requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
openai
better-profanity
requests
authlib
jwt
itsdangerous